Adds option to disable auto-initialization
Introduces an option to disable the automatic initialization of the DiContainer during the build process.

This allows for more control over when singleton services are initialized, especially eager singletons. A new `autoInit` option is added to the `build` method of `DiConfigurator`.

The default behavior remains the same (auto-initialization is enabled).